+++ src/uuseg_word.mli
@@ -10,4 +10,5 @@
 type t
 val create : unit -> t
 val copy : t -> t
+val equal : t -> t -> bool
 val add : t ->  [ `Await | `End | `Uchar of Uchar.t ] -> Uuseg_base.ret